Reorganization of releases file completed
